+++ upload-server/cmd/cli/hooks.go
@@ -2,19 +2,14 @@ package cli
 
 import (
 	"context"
-	"errors"
-	"io"
-	"io/fs"
 	"net"
-	"net/http"
 	"os"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
-	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/cdcgov/data-exchange-upload/upload-server/internal/appconfig"
+	azureloader "github.com/cdcgov/data-exchange-upload/upload-server/internal/loaders/azure"
+	fileloader "github.com/cdcgov/data-exchange-upload/upload-server/internal/loaders/file"
 	"github.com/cdcgov/data-exchange-upload/upload-server/internal/metadata"
-	"github.com/cdcgov/data-exchange-upload/upload-server/internal/metadata/validation"
 	"github.com/cdcgov/data-exchange-upload/upload-server/internal/storeaz"
 	prebuilthooks "github.com/cdcgov/data-exchange-upload/upload-server/pkg/hooks"
 	tusHooks "github.com/tus/tusd/v2/pkg/hooks"
@@ -31,44 +26,11 @@ func GetHookHandler(appConfig appconfig.AppConfig) (tusHooks.HookHandler, error)
 	return PrebuiltHooks(appConfig)
 }
 
-type FileConfigLoader struct {
-	FileSystem fs.FS
-}
-
-func (l *FileConfigLoader) LoadConfig(ctx context.Context, path string) ([]byte, error) {
-
-	file, err := l.FileSystem.Open(path)
-	if err != nil {
-		return nil, errors.Join(err, validation.ErrNotFound)
-	}
-	return io.ReadAll(file)
-}
-
-type AzureConfigLoader struct {
-	Client        *azblob.Client
-	ContainerName string
-}
-
-func (l *AzureConfigLoader) LoadConfig(ctx context.Context, path string) ([]byte, error) {
-	downloadResponse, err := l.Client.DownloadStream(ctx, l.ContainerName, path, nil)
-	if err != nil {
-		var respErr *azcore.ResponseError
-		if errors.As(err, &respErr) {
-			if respErr.StatusCode == http.StatusNotFound {
-				return nil, errors.Join(err, validation.ErrNotFound)
-			}
-		}
-		return nil, err
-	}
-
-	return io.ReadAll(downloadResponse.Body)
-}
-
 func PrebuiltHooks(appConfig appconfig.AppConfig) (tusHooks.HookHandler, error) {
 	handler := &prebuilthooks.PrebuiltHook{}
 
 	preCreateHook := metadata.SenderManifestVerification{
-		Loader: &FileConfigLoader{
+		Loader: &fileloader.FileConfigLoader{
 			FileSystem: os.DirFS(appConfig.UploadConfigPath),
 		},
 		Reporter: &metadata.FileReporter{
@@ -90,7 +52,7 @@ func PrebuiltHooks(appConfig appconfig.AppConfig) (tusHooks.HookHandler, error) 
 		if err != nil {
 			return nil, err
 		}
-		preCreateHook.Loader = &AzureConfigLoader{
+		preCreateHook.Loader = &azureloader.AzureConfigLoader{
 			Client:        client,
 			ContainerName: appConfig.AzureManifestConfigContainer,
 		}
