+++ upload-server/internal/loaders/azure/azure.go
@@ -1,0 +1,32 @@
+package azure
+
+import (
+	"context"
+	"errors"
+	"io"
+	"net/http"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
+	"github.com/cdcgov/data-exchange-upload/upload-server/internal/metadata/validation"
+)
+
+type AzureConfigLoader struct {
+	Client        *azblob.Client
+	ContainerName string
+}
+
+func (l *AzureConfigLoader) LoadConfig(ctx context.Context, path string) ([]byte, error) {
+	downloadResponse, err := l.Client.DownloadStream(ctx, l.ContainerName, path, nil)
+	if err != nil {
+		var respErr *azcore.ResponseError
+		if errors.As(err, &respErr) {
+			if respErr.StatusCode == http.StatusNotFound {
+				return nil, errors.Join(err, validation.ErrNotFound)
+			}
+		}
+		return nil, err
+	}
+
+	return io.ReadAll(downloadResponse.Body)
+}
